+++ tests/BLS12_381.test.js
@@ -1,0 +1,4 @@
+const she = require('../she')
+const { curveTest } = require('./helper')
+
+curveTest(she.BLS12_381, 'BLS12_381')
