+++ tests/BN254.test.js
@@ -1,0 +1,4 @@
+const she = require('../she')
+const { curveTest } = require('./helper')
+
+curveTest(she.BN254, 'BN254')
