+++ tests/BN381_1.test.js
@@ -1,0 +1,4 @@
+const she = require('../she')
+const { curveTest } = require('./helper')
+
+curveTest(she.BN381_1, 'BN381_1')
